Add files via upload
FujiCoin was launched on Saturday, 28th Jun 2014 21:00:00 GMT
FujiCoin is a cryptocurrency that enables instant, near-zero cost payments to anyone in the world. FujiCoin is an open global payment network that is fully decentralized. Mathematics secures the network and empowers individuals to control their own finances. The time required for Fujicoin's transaction confirmation is one tenth of the Bitcoin, it has been greatly speeded up. Also, since the processing capacity of the transaction is 10 times the Bitcoin, there is no worry about the remittance delay. With substantial industry support, trade volume and liquidity, FujiCoin is a proven medium of commerce complementary to Bitcoin and Litecoin.